pyproject.toml
[tool.poetry]
name = "swarm-checkin-post-man"
version = "0.1.0"
description = ""
authors = ["Shunsuke KITADA <shunsuke.kitada.0831@gmail.com>"]
readme = "README.md"
packages = [{ include = "scpm" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
tweepy = "^4.14.0"
fastapi = { extras = ["standard"], version = "^0.115.0" }
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.1.5"
mypy = ">=1.0.0"
pytest = ">=6.0.0"
types-requests = "^2.32.0.20240914"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
